requirements.txt
streamlit>=1.45.0
networkx>=3.4.2
torch>=2.7.0
requests>=2.32.3
numpy>=2.2.5
pandas>=2.2.3
langchain>=0.3.24
langchain-community>=0.3.23
langchain-core>=0.3.56
faiss-cpu>=1.11.0  
sentence-transformers>=4.1.0
rank-bm25>=0.2.2
pydantic>=2.11.4
PyMuPDF>=1.25.5
python-docx>=1.1.2
tqdm>=4.67.1
pypdf>=5.4.0
python-dotenv>=1.1.0
bcrypt>=4.3.0
python-jose>=3.4.0
openai>=1.76.2
jieba>=0.42.1
langchain_ollama>=0.3.2